# public/docs/api-docs.json
{
    "openapi": "3.0.0",
    "info": {
        "title": "Coligo Backend API",
        "description": "API documentation for Coligo - Platform for collaborative transport",
        "contact": {
            "email": "support@coligo.com"
        },
        "license": {
            "name": "MIT",
            "url": "https://opensource.org/licenses/MIT"
        },
        "version": "1.0.0"
    },
    "servers": [
        {
            "url": "http://localhost:8000",
            "description": "Development Server"
        }
    ],
    "paths": {
...
        "/api/matches/suggestions": {
            "get": {
                "tags": [
                    "Matches"
                ],
                "summary": "Get automatic match suggestions",
                "operationId": "getMatchSuggestions",
                "parameters": [
                    {
                        "name": "type",
                        "in": "query",
                        "description": "Type of suggestions: for_shipments, for_routes",
                        "schema": {
                            "type": "string",
                            "enum": [
                                "for_shipments",
                                "for_routes"
                            ]
                        }
                    },
                    {
                        "name": "limit",
                        "in": "query",
                        "description": "Number of suggestions to return",
                        "schema": {
                            "type": "integer",
                            "default": 10,
                            "maximum": 50
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "Suggestions retrieved successfully",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "properties": {
                                        "success": {
                                            "type": "boolean"
                                        },
                                        "data": {
                                            "properties": {
                                                "suggestions": {
                                                    "type": "array",
                                                    "items": {
                                                        "type": "object"
                                                    }
                                                }
                                            },
                                            "type": "object"
                                        }
                                    },
                                    "type": "object"
                                }
                            }
                        }
                    }
                },
                "security": [
                    {
                        "sanctum": []
                    }
                ]
            }
        },
        "/api/matches": {
            "post": {
                "tags": [
                    "Matches"
                ],
                "summary": "Create manual match proposal",
                "operationId": "createMatch",
                "requestBody": {
                    "required": true,
                    "content": {
                        "application/json": {
                            "schema": {
                                "required": [
                                    "shipment_id",
                                    "route_id",
                                    "proposed_price"
                                ],
                                "properties": {
                                    "shipment_id": {
                                        "type": "integer",
                                        "example": 1
                                    },
                                    "route_id": {
                                        "type": "integer",
                                        "example": 1
                                    },
                                    "proposed_price": {
                                        "type": "number",
                                        "format": "float",
                                        "example": 75.5
                                    },
                                    "pickup_datetime": {
                                        "type": "string",
                                        "format": "datetime"
                                    },
                                    "delivery_datetime": {
                                        "type": "string",
                                        "format": "datetime"
                                    },
                                    "transporter_message": {
                                        "type": "string"
                                    }
                                },
                                "type": "object"
                            }
                        }
                    }
                },
                "responses": {
                    "201": {
                        "description": "Match proposal created successfully"
                    },
                    "404": {
                        "description": "Shipment or route not found"
                    },
                    "422": {
                        "description": "Validation errors"
                    }
                },
                "security": [
                    {
                        "sanctum": []
                    }
                ]
            }
        },
        "/api/matches/{id}": {
            "get": {
                "tags": [
                    "Matches"
                ],
                "summary": "Get match details",
                "operationId": "showMatch",
                "parameters": [
                    {
                        "name": "id",
                        "in": "path",
                        "required": true,
                        "schema": {
                            "type": "integer"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "Match retrieved successfully"
                    },
                    "404": {
                        "description": "Match not found"
                    }
                },
                "security": [
                    {
                        "sanctum": []
                    }
                ]
            }
        },
        "/api/matches/{id}/accept": {
            "put": {
                "tags": [
                    "Matches"
                ],
                "summary": "Accept match proposal",
                "operationId": "acceptMatch",
                "parameters": [
                    {
                        "name": "id",
                        "in": "path",
                        "required": true,
                        "schema": {
                            "type": "integer"
                        }
                    }
                ],
                "requestBody": {
                    "content": {
                        "application/json": {
                            "schema": {
                                "properties": {
                                    "final_price": {
                                        "type": "number",
                                        "format": "float"
                                    },
                                    "sender_response": {
                                        "type": "string"
                                    }
                                },
                                "type": "object"
                            }
                        }
                    }
                },
                "responses": {
                    "200": {
                        "description": "Match accepted successfully"
                    }
                },
                "security": [
                    {
                        "sanctum": []
                    }
                ]
            }
        },
        "/api/matches/{id}/reject": {
            "put": {
                "tags": [
                    "Matches"
                ],
                "summary": "Reject match proposal",
                "operationId": "rejectMatch",
                "parameters": [
                    {
                        "name": "id",
                        "in": "path",
                        "required": true,
                        "schema": {
                            "type": "integer"
                        }
                    }
                ],
                "requestBody": {
                    "content": {
                        "application/json": {
                            "schema": {
                                "properties": {
                                    "sender_response": {
                                        "type": "string"
                                    }
                                },
                                "type": "object"
                            }
                        }
                    }
                },
                "responses": {
                    "200": {
                        "description": "Match rejected successfully"
                    }
                },
                "security": [
                    {
                        "sanctum": []
                    }
                ]
            }
        },
        "/api/matches/{id}/complete": {
            "put": {
                "tags": [
                    "Matches"
                ],
                "summary": "Mark match as completed",
                "operationId": "completeMatch",
                "parameters": [
                    {
                        "name": "id",
                        "in": "path",
                        "required": true,
                        "schema": {
                            "type": "integer"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "Match completed successfully"
                    }
                },
                "security": [
                    {
                        "sanctum": []
                    }
                ]
            }
        },
...
    },
    "components": {
        "securitySchemes": {
            "sanctum": {
                "type": "http",
                "description": "Laravel Sanctum token authentication",
                "bearerFormat": "JWT",
                "scheme": "bearer"
            }
        }
    },
...
}